% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.mxSummary.R
\name{ds.mxSummary}
\alias{ds.mxSummary}
\title{Summarizes the results of an mx model optimization}
\usage{
ds.mxSummary(mxObj, datasources = NULL)
}
\arguments{
\item{mxObj}{a character string, the name of an object obtained after optimizing a model via 'ds.mxRun'.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{
the summary of the output of an 'ds.mxRun' command.
}
\description{
This function is similar to base function 'summary'.
}
\details{
to be written
}
\examples{
{

}

}
\author{
Gaye, A.
}
