% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ds.omxGetParameters}
\alias{ds.omxGetParameters}
\title{Return a vector of the free parameters in the model.}
\usage{
ds.omxGetParameters(model, indep = FALSE, free = c(TRUE, FALSE, NA),
  fetch = c("values", "free", "lbound", "ubound", "all"),
  datasources = NULL)
}
\arguments{
\item{model}{a MxModel object}

\item{indep}{fetch parameters from independent submodels}

\item{free}{fetch either free parameters (TRUE), or fixed parameters or both types. Default value is TRUE.}

\item{fetch}{which attribute of the parameters to fetch. Default choice is ‘values’}
}
\value{
a vector of free (by default) parameters in the model
}
\description{
This function is similar to OpenMx function \code{omxGetParameters}.
}
\details{


See details of the OpenMx function 'omxGetParameters' in the package \code{OpenMx}.
}
\examples{
{

}
}
\author{
Gaye, A., Kutschke, J.
}
\references{
Steven M. Boker, Michael C. Neale, Hermine H. Maes, Michael J. Wilde, Michael Spiegel, Timothy R. Brick,
Jeffrey Spies, Ryne Estabrook, Sarah Kenny, Timothy C., Bates, Paras Mehta, and John Fox. (2011)
OpenMx: An Open Source Extended Structural Equation Modeling Framework. Psychometrika.
Steven M. Boker, Michael C. Neale, Hermine H. Maes, Michael J. Wilde, Michael Spiegel, Timothy R. Brick, Ryne Estabrook,
Timothy C. Bates, Paras Mehta, Timo von Oertzen, Ross J. Gore, Michael D. Hunter, Daniel C. Hackett, Julian Karch and
Andreas M. Brandmaier. (2012) OpenMx 1.3 User Guide.
}

