% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ds.omxSetParameters}
\alias{ds.omxSetParameters}
\title{Assign model parameters}
\usage{
ds.omxSetParameters(model, labels, free = NULL, values = NULL,
  newlabels = NULL, lbound = NULL, ubound = NULL, indep = FALSE,
  strict = TRUE, name = NULL, newobj = "new_modified_mxModel",
  datasources = NULL)
}
\arguments{
\item{model}{a MxModel object}

\item{labels}{a character vector of target parameter names}

\item{free}{a boolean vector of parameter free/fixed designations}

\item{values}{a numeric vector of parameter values}

\item{newlabels}{a character vector of new parameter names}

\item{lbound}{a numeric vector of lower bound values}

\item{ubound}{a numeric vector of upper bound values}

\item{indep}{boolean. set parameters in independent submodels}

\item{strict}{boolean. if TRUE then throw an error when a label does not appear in the model}

\item{name}{character string. (optional) a new name for the model}
}
\value{
a new modified mxModel
}
\description{
This function is similar to OpenMx function \code{omxSetParameters}.
}
\details{


See details of the OpenMx function 'omxSetParameters' in the package \code{OpenMx}.
}
\examples{
{

}
}
\author{
Gaye, A., Kutschke, J.
}
\references{
Steven M. Boker, Michael C. Neale, Hermine H. Maes, Michael J. Wilde, Michael Spiegel, Timothy R. Brick,
Jeffrey Spies, Ryne Estabrook, Sarah Kenny, Timothy C., Bates, Paras Mehta, and John Fox. (2011)
OpenMx: An Open Source Extended Structural Equation Modeling Framework. Psychometrika.
Steven M. Boker, Michael C. Neale, Hermine H. Maes, Michael J. Wilde, Michael Spiegel, Timothy R. Brick, Ryne Estabrook,
Timothy C. Bates, Paras Mehta, Timo von Oertzen, Ross J. Gore, Michael D. Hunter, Daniel C. Hackett, Julian Karch and
Andreas M. Brandmaier. (2012) OpenMx 1.3 User Guide.
}

