\name{getCall}
\alias{getCall}
\title{deals with expressions with that include symbols}
\usage{
getCall(input, objnames, attnames)
}
\arguments{
  \item{input}{an expression for 'ds.mxAlgebra'}

  \item{objnames}{an optional character string indicating
  the name of the object.}

  \item{attnames}{a list, the dimnames attribute for the
  algebra}
}
\value{
an object of type call
}
\description{
This is an internal function.
}
\details{


Not required
}
\keyword{internal}

