% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.mxAlgebra.R
\name{ds.mxAlgebra}
\alias{ds.mxAlgebra}
\title{Creates a new MxAlgebra object}
\usage{
ds.mxAlgebra(expression = NULL, name = NA, dimnames = NA,
  newobj = "mxAlgebra_output", datasources = NULL)
}
\arguments{
\item{expression}{an R expression of OpenMx-supported matrix operators and matrix functions.}

\item{name}{an optional character string indicating the name of the object.}

\item{dimnames}{a list of character strings, the dimnames attribute for the algebra}

\item{newobj}{the name of the new variable. If this argument is set to NULL, the name of the new
object is "mxAlgebra_output".}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\value{
an object of type 'mxAlgebra'
}
\description{
This function is similar to OpenMx function \code{mxAlgebra}.
}
\details{
See details of the OpenMx function 'mxAlbegra
}
\examples{
{

}
}
\references{
Steven M. Boker, Michael C. Neale, Hermine H. Maes, Michael J. Wilde, Michael Spiegel, Timothy R. Brick,
Jeffrey Spies, Ryne Estabrook, Sarah Kenny, Timothy C., Bates, Paras Mehta, and John Fox. (2011)
OpenMx: An Open Source Extended Structural Equation Modeling Framework. Psychometrika.
Steven M. Boker, Michael C. Neale, Hermine H. Maes, Michael J. Wilde, Michael Spiegel, Timothy R. Brick, Ryne Estabrook,
Timothy C. Bates, Paras Mehta, Timo von Oertzen, Ross J. Gore, Michael D. Hunter, Daniel C. Hackett, Julian Karch and
Andreas M. Brandmaier. (2012) OpenMx 1.3 User Guide.
}
\author{
Gaye, A.
}
